+++ src/components/product/ProductSummary/ProductSummary.tsx
@@ -3,6 +3,7 @@ import { graphql, Link } from 'gatsby'
 import { GatsbyImage } from 'gatsby-plugin-image'
 import React from 'react'
 import imagesConf from 'src/images/config'
+import { useBuyButton } from 'src/sdk/cart/useBuyButton'
 
 import type { ProductSummary_ProductFragment } from './__generated__/ProductSummary_product.graphql'
 
@@ -11,18 +12,32 @@ interface Props {
 }
 
 function ProductSummary({ product }: Props) {
-  const { imageUrl: src, imageText: alt } =
-    product.items?.[0]?.images?.[0] ?? {}
+  const { images, sellers } = product.items?.[0] ?? {}
+  const { imageUrl: src, imageText: alt } = images?.[0] ?? {}
+  const offer = sellers?.[0]?.commercialOffer
 
   const image = useThumborImageData({
     baseUrl: src ?? '',
     ...imagesConf['product.summary'],
   })
 
+  const buyProps = useBuyButton(
+    offer && {
+      id: product.id!,
+      price: offer.spotPrice!,
+      listPrice: offer.listPrice!,
+      quantity: {
+        selling: 1,
+        gift: 0,
+      },
+    }
+  )
+
   return (
     <Link to={`/${product.slug}/p`}>
       <GatsbyImage image={image} alt={alt ?? ''} />
       <div>{product.productName}</div>
+      <button {...buyProps}>Add to cart</button>
     </Link>
   )
 }
@@ -30,13 +45,19 @@ function ProductSummary({ product }: Props) {
 export const fragment = graphql`
   fragment ProductSummary_product on VTEX_Product {
     slug: linkText
-    productId
+    id: productId
     productName
 
     items {
       images {
         imageUrl
         imageText
+      }
+      sellers {
+        commercialOffer: commertialOffer {
+          spotPrice
+          listPrice: ListPrice
+        }
       }
     }
   }
