+++ src/components/sections/ProductDetails/ProductDetails.tsx
@@ -1,37 +1,26 @@
-import React, { useCallback } from 'react'
-import { useCart, useGlobalUIState } from '@vtex/store-sdk'
+import { useThumborImageData } from '@vtex/gatsby-plugin-thumbor'
 import { graphql } from 'gatsby'
-import type { CartItem } from '@vtex/store-sdk'
+import { GatsbyImage } from 'gatsby-plugin-image'
+import React from 'react'
+import imagesConf from 'src/images/config'
+import { useBuyButton } from 'src/sdk/cart/useBuyButton'
 import type { ProductDetailsFragment_ProductFragment } from 'src/views/product/__generated__/ProductViewFragment_product.graphql'
 
 interface Props {
   product: ProductDetailsFragment_ProductFragment
 }
 
-const useBuyButton = (item: CartItem | null | undefined) => {
-  const { addItem } = useCart()
-  const { openMinicart } = useGlobalUIState()
-
-  const onClick = useCallback(
-    (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-      e.preventDefault()
-
-      if (!item) {
-        return
-      }
-
-      addItem(item)
-      openMinicart()
-    },
-    [item, addItem, openMinicart]
-  )
-
-  return { onClick }
-}
-
 function ProductDetails({ product }: Props) {
-  const offer = product.items?.[0]?.sellers?.[0]?.commercialOffer
-  const btnProps = useBuyButton(
+  const { images, sellers } = product.items?.[0] ?? {}
+  const { imageUrl: src, imageText: alt } = images?.[0] ?? {}
+  const offer = sellers?.[0]?.commercialOffer
+
+  const image = useThumborImageData({
+    baseUrl: src ?? '',
+    ...imagesConf['product.details'],
+  })
+
+  const buyProps = useBuyButton(
     offer && {
       id: product.id!,
       price: offer.spotPrice!,
@@ -46,7 +35,8 @@ function ProductDetails({ product }: Props) {
   return (
     <>
       <h1>{product.productName}</h1>
-      <button {...btnProps}>Buy Now</button>
+      <GatsbyImage image={image} alt={alt ?? ''} />
+      <button {...buyProps}>Add to cart</button>
     </>
   )
 }
@@ -57,6 +47,10 @@ export const fragment = graphql`
     productName
 
     items {
+      images {
+        imageUrl
+        imageText
+      }
       sellers {
         commercialOffer: commertialOffer {
           spotPrice
